# Turbopack: optimize NftJsonAsset