--- a/crates/next-api/src/nft_json.rs
+++ b/crates/next-api/src/nft_json.rs
@@ -3,7 +3,7 @@ use std::collections::BTreeSet;
 use anyhow::{bail, Result};
 use serde_json::json;
 use turbo_rcstr::RcStr;
-use turbo_tasks::{ResolvedVc, ValueToString, Vc};
+use turbo_tasks::{ResolvedVc, Vc};
 use turbo_tasks_fs::{File, FileSystem, FileSystemPath};
 use turbopack_core::{
     asset::{Asset, AssetContent},
@@ -50,101 +50,10 @@ impl NftJsonAsset {
     }
 
     #[turbo_tasks::function]
-    fn output_root(&self) -> Vc<FileSystemPath> {
-        self.project.output_fs().root()
-    }
-
-    #[turbo_tasks::function]
-    fn project_root(&self) -> Vc<FileSystemPath> {
-        self.project.project_fs().root()
-    }
-
-    #[turbo_tasks::function]
-    fn client_root(&self) -> Vc<FileSystemPath> {
-        self.project.client_fs().root()
-    }
-
-    #[turbo_tasks::function]
-    fn project_path(&self) -> Vc<FileSystemPath> {
-        self.project.project_path()
-    }
-
-    #[turbo_tasks::function]
     async fn dist_dir(&self) -> Result<Vc<RcStr>> {
         Ok(Vc::cell(
             format!("/{}/", self.project.dist_dir().await?).into(),
         ))
-    }
-}
-
-#[turbo_tasks::value(transparent)]
-pub struct OutputSpecifier(Option<RcStr>);
-
-#[turbo_tasks::value_impl]
-impl NftJsonAsset {
-    #[turbo_tasks::function]
-    async fn ident_folder(self: Vc<Self>) -> Vc<FileSystemPath> {
-        self.ident().path().parent()
-    }
-
-    #[turbo_tasks::function]
-    async fn ident_folder_in_project_fs(self: Vc<Self>) -> Result<Vc<FileSystemPath>> {
-        Ok(self
-            .project_path()
-            .join(self.ident_folder().await?.path.clone()))
-    }
-
-    #[turbo_tasks::function]
-    async fn ident_folder_in_client_fs(self: Vc<Self>) -> Result<Vc<FileSystemPath>> {
-        Ok(self
-            .client_root()
-            .join(self.ident_folder().await?.path.clone()))
-    }
-
-    #[turbo_tasks::function]
-    async fn get_output_specifier(
-        self: Vc<Self>,
-        path: Vc<FileSystemPath>,
-    ) -> Result<Vc<OutputSpecifier>> {
-        let path_ref = path.await?;
-
-        // include assets in the outputs such as referenced chunks
-        if path_ref.is_inside_ref(&*(self.output_root().await?)) {
-            return Ok(Vc::cell(Some(
-                self.ident_folder()
-                    .await?
-                    .get_relative_path_to(&path_ref)
-                    .unwrap(),
-            )));
-        }
-
-        // include assets in the project root such as images and traced references (externals)
-        if path_ref.is_inside_ref(&*(self.project_root().await?)) {
-            return Ok(Vc::cell(Some(
-                self.ident_folder_in_project_fs()
-                    .await?
-                    .get_relative_path_to(&path_ref)
-                    .unwrap(),
-            )));
-        }
-
-        // assets that are needed on the client side such as fonts and icons
-        if path_ref.is_inside_ref(&*(self.client_root().await?)) {
-            return Ok(Vc::cell(Some(
-                self.ident_folder_in_client_fs()
-                    .await?
-                    .get_relative_path_to(&path_ref)
-                    .unwrap()
-                    .replace("/_next/", &self.dist_dir().await?)
-                    .into(),
-            )));
-        }
-
-        // Make this an error for now, this should effectively be unreachable
-        bail!(
-            "NftJsonAsset: cannot handle filepath {}",
-            path.to_string().await?
-        );
     }
 }
 
@@ -161,12 +70,64 @@ impl OutputAsset for NftJsonAsset {
     }
 }
 
+#[turbo_tasks::value(transparent)]
+pub struct OutputSpecifier(Option<RcStr>);
+
+fn get_output_specifier(
+    path_ref: &FileSystemPath,
+    ident_folder: &FileSystemPath,
+    ident_folder_in_project_fs: &FileSystemPath,
+    ident_folder_in_client_fs: &FileSystemPath,
+    output_root: &FileSystemPath,
+    project_root: &FileSystemPath,
+    client_root: &FileSystemPath,
+    dist_dir: &RcStr,
+) -> Result<RcStr> {
+    // include assets in the outputs such as referenced chunks
+    if path_ref.is_inside_ref(output_root) {
+        return Ok(ident_folder.get_relative_path_to(path_ref).unwrap());
+    }
+
+    // include assets in the project root such as images and traced references (externals)
+    if path_ref.is_inside_ref(project_root) {
+        return Ok(ident_folder_in_project_fs
+            .get_relative_path_to(path_ref)
+            .unwrap());
+    }
+
+    // assets that are needed on the client side such as fonts and icons
+    if path_ref.is_inside_ref(client_root) {
+        return Ok(ident_folder_in_client_fs
+            .get_relative_path_to(path_ref)
+            .unwrap()
+            .replace("/_next/", dist_dir)
+            .into());
+    }
+
+    // Make this an error for now, this should effectively be unreachable
+    bail!("NftJsonAsset: cannot handle filepath {}", path_ref);
+}
+
 #[turbo_tasks::value_impl]
 impl Asset for NftJsonAsset {
     #[turbo_tasks::function]
     async fn content(self: Vc<Self>) -> Result<Vc<AssetContent>> {
         let this = &*self.await?;
         let mut result = BTreeSet::new();
+
+        let output_root_ref = this.project.output_fs().root().await?;
+        let project_root_ref = this.project.project_fs().root().await?;
+        let client_root = this.project.client_fs().root();
+        let client_root_ref = client_root.await?;
+        let dist_dir = self.dist_dir().await?;
+
+        let ident_folder = self.ident().path().parent().await?;
+        let ident_folder_in_project_fs = this
+            .project
+            .project_path()
+            .join(ident_folder.path.clone())
+            .await?;
+        let ident_folder_in_client_fs = client_root.join(ident_folder.path.clone()).await?;
 
         let chunk = this.chunk;
         let entries = this
@@ -176,20 +137,26 @@ impl Asset for NftJsonAsset {
             .chain(std::iter::once(chunk))
             .collect();
         for referenced_chunk in all_assets_from_entries(Vc::cell(entries)).await? {
-            if referenced_chunk.ident().path().await?.extension_ref() == Some("map") {
-                continue;
-            }
-
             if chunk.eq(referenced_chunk) {
                 continue;
             }
 
-            let specifier = self
-                .get_output_specifier(referenced_chunk.ident().path())
-                .await?;
-            if let Some(specifier) = &*specifier {
-                result.insert(specifier.clone());
+            let referenced_chunk_path = referenced_chunk.ident().path().await?;
+            if referenced_chunk_path.extension_ref() == Some("map") {
+                continue;
             }
+
+            let specifier = get_output_specifier(
+                &referenced_chunk_path,
+                &ident_folder,
+                &ident_folder_in_project_fs,
+                &ident_folder_in_client_fs,
+                &output_root_ref,
+                &project_root_ref,
+                &client_root_ref,
+                &dist_dir,
+            )?;
+            result.insert(specifier);
         }
 
         let json = json!({