before:
aggregated NftJsonAsset::content 2.13s ⌀22.38ms (24.10s ⌀253.66ms) (11.3x)
aggregated NftJsonAsset::content [1032.29MB ⌀10.87MB] ([6.37GB ⌀67.10MB])
total build: 266.84s user 58.97s system 748% cpu 43.505 total
total memory: 16.74gb
![Bildschirmfoto 2025-01-20 um 14 24 10](https://github.com/user-attachments/assets/d59ce80c-8cdd-4e16-b936-a046ec1948f3)

after:
aggregated NftJsonAsset::content 480.67ms ⌀5.06ms (4.97s ⌀52.31ms) (10.3x)
aggregated NftJsonAsset::content [119.08MB ⌀1253.44kB] ([1819.16MB ⌀19.15MB])
total build: 255.72s user 55.15s system 735% cpu 42.241 total
total memory: 15.84bg
![Bildschirmfoto 2025-01-20 um 14 30 38](https://github.com/user-attachments/assets/4546a838-c3d4-4fe6-868f-10a99fd9004f)